requirements.txt
aiohttp
configparser
future


google-genai

# Additional Packages for export Functionality
pandas>=2.0.0

# Extra libraries are installed when needed using pipmaster

pipmaster
pydantic
python-dotenv

# Unicode Collation Algorithm for proper Chinese sorting
pyuca

setuptools
tenacity

# LLM packages
tiktoken
xlsxwriter>=3.1.0
sentence-transformers

# monitoring
fastapi 
uvicorn
opik



# graph
neo4j 
pandas

